Вся логика перенесена в модуь libvbukv
