+++ src/libvbukv/file.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use std::{
     fs::File,
     io::{BufRead, BufReader},
@@ -11,6 +14,3 @@ pub fn words_from_file(filename: impl AsRef<Path>) -> Vec<String> {
         .map(|l| l.expect("Could not parse line"))
         .collect()
 }
-
-#[cfg(test)]
-mod tests;
